noheader: place branch combo on tabs if no header
noheader=1 stops the static page header from being emitted by
cgit, but along with that, the small form that lets the user
change branch context on the page is also lost.

This isn't actually static since it contains a dynamic list of
branches; it can't be reproduced on the user's external header
static html.  So it seems it doesn't belong to the set of header
things that should be disabled.

This patch relocates the branch selection combo and
form on the left of the tabs line if noheader=1.  It doesn't
change anything if noheader is not set.

Signed-off-by: Andy Green <andy@warmcat.com>
Reviewed-by: John Keeping <john@keeping.me.uk>