cgit.js: line range highlight: make responsive to url changes
Browser default interpretation of the # part of the URL is
to try to match it to an element ID one time at page load.

Subsequent modifications in the URL bar are ignored, even if
you hit enter there.

This patch makes the line range highlight able to clean up
after itself and be reapplied, and has the highlight function
listen for hashchange events and re-interpret the # part
when they occur.

Now if you edit the URL and press enter, any existing highlight
is removed, the # part reinterpreted and the new highlight
applied automatically.

This is particularly useful if you edit the # link with the
intention to show a range before copying the URL.

Clicking on the line number, which changes the URL bar to
have a corresponding #n part, also triggers hashchange and
makes the clicked line be highlit immediately.

Signed-off-by: Andy Green <andy@warmcat.com>