line-range-highlight: burger menu and popup menu
Clicking on the line numbers to control the highlight,
or visiting a #URL link, causes a burger menu to
appear on the left of either the top of the range
or the bottom if that was last clicked.

Clicking this brings up a popup menu with

Copy Lines (implemented in next patch)
Copy Link (copies complete #URL to clipboard)
View / Remove Blame (changes to view to have or not have blame)

Clicking outside the popup menu clears it, but the burger
menu is sticky.

This is an entirely clientside implementation in
cgit.css and cgit.js only.  If JS disabled at server
or at client, it cannot provide this functionality and
operates with one-line browser URLs as before.

Tested on Linux Chrome 67 + Firefox 60.

Signed-off-by: Andy Green <andy@warmcat.com>